backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
python-dotenv==1.0.1
textblob==0.18.0
faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy==1.26.4
scikit-learn==1.5.1
beautifulsoup4==4.12.3
httpx==0.27.2
pydantic==2.9.1
